Misc tweaks

+++ src/Common/Utils/MathUtils.cs
@@ -216,9 +216,9 @@ namespace Common.Utils
         /// <summary>
         /// Maps a 0-255 byte value to a 0°-180° angle in radians.
         /// </summary>
-        public static double ByteToAngle(this byte b)
+        public static float ByteToAngle(this byte b)
         {
-            return b / 255.0 * Math.PI;
+            return (float)(b / 255.0 * Math.PI);
         }
 
         /// <summary>
@@ -226,7 +226,7 @@ namespace Common.Utils
         /// </summary>
         public static Vector4 ByteToAngle(this ByteVector4 vector)
         {
-            return new Vector4((float)vector.X.ByteToAngle(), (float)vector.Y.ByteToAngle(), (float)vector.Z.ByteToAngle(), (float)vector.W.ByteToAngle());
+            return new Vector4(vector.X.ByteToAngle(), vector.Y.ByteToAngle(), vector.Z.ByteToAngle(), vector.W.ByteToAngle());
         }
         #endregion
